requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

matplotlib==3.2.1
numpy==1.18.4
opencv_python_headless==4.6.0.66
pandas==1.3.5
Pillow==9.0.1
scikit_image==0.19.3
scikit_learn==0.22.2.post1
scipy==1.4.1
tabulate==0.9.0
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.64.1
ptflops==0.7
tensorboard==2.11.0